Add with-cache example to the readme.
Per issue #2 the documentation both here and in the org registry listing is pretty bad.

I.e. it took ~30 minutes of failed builds + googling just to get ^ which is essentially the orb-version of the "getting started with node" snippet from the "set up job" flow.

Granted, the orb registry has detailed documentation of the step definitions, but no concrete examples that show things like the `node/` prefix being required for orb-defined commands.