Advanced conditionals [semver:major] (#38)
New features:

Can override the package installation command with a custom command <<parameters.override-ci-command>>.
This could also be used if additional flags are needed for the package installation command.
Caching is now highly opinionated but fully automatic.
The cache key parameter is now removed. package-lock.json will be used for NPM, and yarn.lock will be used for YARN. It is possible to add this parameter back if needed but will introduce a lot of duplicated config in the orb source.
The caching path is also decided automatically based on the package manager, however it is still parameterized and can still be overwritten if needed.
By default, $HOME/.npm is used for NPM, and the node_modules folder for YARN.
Removed:

override-ci boolean parameter no longer required, and so removed. This is however breaking if a user had previously used this parameter.
cache-key parameter removed. Now decided automatically.
Highly opinionated but automatically selects the right type of file depending on the package manager.
The down-side to making this parameter available to override is we need to duplicate our logic in the orb for each combination of parameters. Since we left in the ability to specify the cache-path but also specify one by default, that's two scenarios, adding the cache-key would mean another two duplications. Though it is possible to do this if needed
Issues addressed:

#35
#29
#25
Other:

Updated code owners
Test for overriding CI command and using YARN
Improved cache testing